+++ billinghelper/src/main/kotlin/com/vojtkovszky/billinghelper/BillingBuilderConfig.kt
@@ -17,11 +17,13 @@ import com.android.billingclient.api.UserChoiceBillingListener
  * @param userChoiceBillingListener build client with [BillingClient.Builder.enableUserChoiceBilling] and provide given listener.
  * For more info see [https://developer.android.com/reference/com/android/billingclient/api/UserChoiceBillingListener]
  * and [https://support.google.com/googleplay/android-developer/answer/13821247]
+ * @param enableAutoServiceReconnection see [https://developer.android.com/google/play/billing/integrate#automatic-service-reconnection]
  */
 class BillingBuilderConfig(
     val enableAlternativeBillingOnly: Boolean = false,
     val enableExternalOffer: Boolean = false,
     val enablePendingPurchasesOneTimeProducts: Boolean = true,
     val enablePendingPurchasesPrepaidPlans: Boolean = true,
-    val userChoiceBillingListener: UserChoiceBillingListener? = null
+    val userChoiceBillingListener: UserChoiceBillingListener? = null,
+    val enableAutoServiceReconnection: Boolean = true,
 )