+++ billinghelper/src/main/kotlin/com/vojtkovszky/billinghelper/BillingHelper.kt
@@ -6,9 +6,6 @@ import android.os.Handler
 import android.os.Looper
 import android.util.Log
 import com.android.billingclient.api.*
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 /**
  * Construct helper. By default, connection will be initialized immediately with product details and
@@ -23,7 +20,6 @@ import kotlinx.coroutines.launch
  * @param billingBuilderConfig additional configuration used when building [BillingClient]. Default covers most common use cases.
  * @param queryProductDetailsOnConnected set whether [initQueryProductDetails] should be called automatically right after client connects (when [initClientConnection] succeeds).
  * @param queryOwnedPurchasesOnConnected set whether [initQueryOwnedPurchases] should be called automatically right after client connects (when [initClientConnection] succeeds).
- * @param queryPurchaseHistoryRecordsOnConnected set whether [initQueryPurchaseHistoryRecords] should be called automatically right after client connects (when [initClientConnection] succeeds).
  * @param autoAcknowledgePurchases All purchases require acknowledgement.
  * By default, this is handled automatically every time state of purchases changes.
  * If set to [Boolean.false], make sure [acknowledgePurchases] is used manually.
@@ -40,7 +36,6 @@ class BillingHelper(
     private var billingBuilderConfig: BillingBuilderConfig = BillingBuilderConfig(),
     var queryProductDetailsOnConnected: Boolean = true,
     var queryOwnedPurchasesOnConnected: Boolean = true,
-    var queryPurchaseHistoryRecordsOnConnected: Boolean = false,
     var autoAcknowledgePurchases: Boolean = true,
     var enableLogging: Boolean = false,
     billingListener: BillingListener? = null,
@@ -51,8 +46,6 @@ class BillingHelper(
 
     // represents list of all currently owned purchases
     private val purchases = mutableListOf<Purchase>()
-    // represents purchase history records
-    private val purchaseHistoryRecords = mutableListOf<PurchaseHistoryRecord>()
     // represents details of all available product details
     private val productDetailsList = mutableListOf<ProductDetails>()
     // callback listeners
@@ -163,6 +156,9 @@ class BillingHelper(
                 billingBuilderConfig.userChoiceBillingListener?.let {
                     enableUserChoiceBilling(it)
                 }
+                if (billingBuilderConfig.enableAutoServiceReconnection) {
+                    enableAutoServiceReconnection()
+                }
             }
             .setListener { billingResult, purchases -> // PurchasesUpdatedListener
                 val billingEvent = when {
@@ -191,8 +187,7 @@ class BillingHelper(
         if (startConnectionImmediately) {
             initClientConnection(
                 queryProductDetailsOnConnected = queryProductDetailsOnConnected,
-                queryOwnedPurchasesOnConnected = queryOwnedPurchasesOnConnected,
-                queryPurchaseHistoryRecordsOnConnected = queryPurchaseHistoryRecordsOnConnected
+                queryOwnedPurchasesOnConnected = queryOwnedPurchasesOnConnected
             )
         }
     }
@@ -259,15 +254,6 @@ class BillingHelper(
             if (isPurchased(productName)) return true
         }
         return false
-    }
-
-    /**
-     * Will return all [PurchaseHistoryRecord] containing a given [productName] or empty list if no match found.
-     * Note that you need to query for owned purchases using [initQueryPurchaseHistoryRecords] or complete a
-     * purchase in order for this to return non-empty.
-     */
-    fun getPurchaseHistoryRecords(productName: String): List<PurchaseHistoryRecord> {
-        return purchaseHistoryRecords.filter { it.products.contains(productName) }
     }
 
     /**
@@ -345,12 +331,9 @@ class BillingHelper(
      * right after client connects. Defaults to [BillingHelper.queryProductDetailsOnConnected]
      * @param queryOwnedPurchasesOnConnected set whether [initQueryOwnedPurchases] should be called
      * right after client connects. Defaults to [BillingHelper.queryOwnedPurchasesOnConnected]
-     * @param queryPurchaseHistoryRecordsOnConnected set whether [initQueryPurchaseHistoryRecords] should be called
-     * right after client connects. Defaults to [BillingHelper.queryPurchaseHistoryRecordsOnConnected]
      */
     fun initClientConnection(queryProductDetailsOnConnected: Boolean = this.queryProductDetailsOnConnected,
-                             queryOwnedPurchasesOnConnected: Boolean = this.queryOwnedPurchasesOnConnected,
-                             queryPurchaseHistoryRecordsOnConnected: Boolean = this.queryPurchaseHistoryRecordsOnConnected) {
+                             queryOwnedPurchasesOnConnected: Boolean = this.queryOwnedPurchasesOnConnected) {
         if (billingClient.isReady) {
             invokeListener(BillingEvent.BILLING_CONNECTED, "BillingClient already connected, skipping.")
             return
@@ -378,10 +361,6 @@ class BillingHelper(
                     if (queryOwnedPurchasesOnConnected) {
                         initQueryOwnedPurchases()
                     }
-                    // query for purchase history records
-                    if (queryPurchaseHistoryRecordsOnConnected) {
-                        initQueryPurchaseHistoryRecords()
-                    }
                 }
             }
             override fun onBillingServiceDisconnected() {
@@ -401,18 +380,6 @@ class BillingHelper(
      */
     fun initQueryOwnedPurchases() {
         initQueryOwnedPurchasesForTypes(getAvailableTypes(), 0, mutableListOf())
-    }
-
-    /**
-     * Initialize query for [PurchaseHistoryRecord]s listed for this app.
-     * Will query for both in-app purchases and subscriptions.
-     * Result will be returned using [billingListeners]
-     *
-     * Note if queryPurchaseHistoryRecords=true in the helper constructor, method gets called
-     * automatically when client connects (See [BillingHelper] constructor for more info).
-     */
-    fun initQueryPurchaseHistoryRecords() {
-        initQueryPurchaseHistoryRecordsForTypes(getAvailableTypes(), 0, mutableListOf())
     }
 
     /**
@@ -551,56 +518,6 @@ class BillingHelper(
         }
     }
 
-    // allows to call queryPurchaseHistory recursively on all types
-    private fun initQueryPurchaseHistoryRecordsForTypes(
-        types: List<String>,
-        currentTypeIndex: Int,
-        resultingList: MutableList<PurchaseHistoryRecord>
-    ) {
-        if (currentTypeIndex == types.size) {
-            purchaseHistoryRecords.clear()
-            purchaseHistoryRecords.addAll(resultingList)
-            purchasesQueried = true
-            invokeListener(
-                event = BillingEvent.QUERY_PURCHASES_HISTORY_RECORDS_COMPLETE,
-                message = resultingList.toString(),
-                responseCode = null
-            )
-            return
-        }
-
-        val currentType = types[currentTypeIndex]
-        CoroutineScope(Dispatchers.IO).launch {
-            try {
-                val purchaseHistoryResult = billingClient.queryPurchaseHistory(
-                    QueryPurchaseHistoryParams.newBuilder().setProductType(currentType).build()
-                )
-
-                if (purchaseHistoryResult.billingResult.isResponseOk()) {
-                    purchaseHistoryResult.purchaseHistoryRecordList?.let {
-                        resultingList.addAll(it)
-                    }
-                    initQueryPurchaseHistoryRecordsForTypes(types,
-                        currentTypeIndex + 1,
-                        resultingList
-                    )
-                } else {
-                    invokeListener(
-                        event = BillingEvent.QUERY_PURCHASES_HISTORY_RECORDS_FAILED,
-                        message = purchaseHistoryResult.billingResult.debugMessage,
-                        responseCode = purchaseHistoryResult.billingResult.responseCode
-                    )
-                }
-            } catch (e: Exception) {
-                invokeListener(
-                    event = BillingEvent.QUERY_PURCHASES_HISTORY_RECORDS_FAILED,
-                    message = e.message ?: "Unknown error",
-                    responseCode = BillingClient.BillingResponseCode.ERROR
-                )
-            }
-        }
-    }
-
     // allows to call queryProductDetailsAsync recursively on all types
     private fun initQueryProductDetailsForTypes(
         types: List<String>,
@@ -635,9 +552,9 @@ class BillingHelper(
             }
 
             val params = QueryProductDetailsParams.newBuilder().setProductList(products)
-            billingClient.queryProductDetailsAsync(params.build()) { queryResult, productDetailsList ->
+            billingClient.queryProductDetailsAsync(params.build()) { queryResult: BillingResult, queryProductDetailsListResult: QueryProductDetailsResult ->
                 if (queryResult.isResponseOk()) {
-                    resultingList.addAll(productDetailsList)
+                    resultingList.addAll(queryProductDetailsListResult.productDetailsList)
                     initQueryProductDetailsForTypes(types, currentTypeIndex+1, resultingList)
                 } else {
                     invokeListener(